backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
sentence-transformers==2.2.2
numpy
scikit-learn
